make this compile on non-windows
